Started migrating from all files permission to MediaStore API

+++ lib/utils/storage_utils.dart
@@ -1,29 +1,49 @@
+// ignore_for_file: avoid_slow_async_io
+
 import 'dart:io' as io;
 
+import 'package:device_info_plus/device_info_plus.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:media_store_plus/media_store_plus.dart';
 import 'package:path_provider/path_provider.dart';
 
+import 'constants.dart';
+import 'custom_dialog.dart';
 import 'shared_preferences_util.dart';
 import 'utils.dart';
 
 class StorageUtils {
   /// Create application folder in internal storage
   static Future<void> createFolder() async {
-    try {
-      final hasStoragePerms = await Utils.requestStoragePermissions();
-      if (hasStoragePerms == false) {
-        // Get.snackbar(
-        //   'Oh no!',
-        //   'Not all permissions were granted. Some app features may not work properly',
-        // );
-        Utils.logError('[StorageUtils] - ' +
-            'Looks like some permissions were not granted');
-      }
+    // Set internal appDirectoryPath
+    final io.Directory internalDirectoryPath =
+        await getApplicationDocumentsDirectory();
+    SharedPrefsUtil.putString(
+      'internalDirectoryPath',
+      internalDirectoryPath.path,
+    );
 
-      // Set internal appDirectoryPath
-      final io.Directory internalDirectoryPath =
-          await getApplicationDocumentsDirectory();
-      SharedPrefsUtil.putString(
-          'internalDirectoryPath', internalDirectoryPath.path);
+    // Set current log file path
+    SharedPrefsUtil.putString('currentLogFile', Utils.getNewLogFilename());
+
+    // Get android sdk info
+    final androidDeviceInfo = await DeviceInfoPlugin().androidInfo;
+    final sdkVersion = androidDeviceInfo.version.sdkInt;
+
+    // We have set this otherwise it throws AppFolderNotSetException
+    MediaStore.appFolder = 'OneSecondDiary';
+
+    try {
+      Utils.logInfo('[App Started] - ' + 'Log file created');
+
+      final hasStoragePerms = await Utils.requestStoragePermissions(
+        sdkVersion: sdkVersion,
+      );
+      if (hasStoragePerms == false) {
+        Utils.logError('[StorageUtils] - ' +
+            'Some storage permissions were not granted for sdk version $sdkVersion');
+      }
 
       io.Directory? appDirectory;
       io.Directory? moviesDirectory;
@@ -33,7 +53,7 @@ class StorageUtils {
       String moviesPath = SharedPrefsUtil.getString('moviesPath');
 
       // If it is not stored, dive into the device folders and store it properly
-      if (appPath == '' || moviesPath == '') {
+      if (!appPath.contains('DCIM')) {
         String rootPath = '';
         appDirectory = await getExternalStorageDirectory();
 
@@ -48,10 +68,10 @@ class StorageUtils {
         }
 
         // Storing appPath
-        appPath = '$rootPath/OneSecondDiary/';
+        appPath = '$rootPath/DCIM/OneSecondDiary/';
         SharedPrefsUtil.putString('appPath', appPath);
         // Storing moviesPath
-        moviesPath = '$rootPath/OSD-Movies/';
+        moviesPath = '$rootPath/DCIM/OneSecondDiary/Movies/';
         SharedPrefsUtil.putString('moviesPath', moviesPath);
       }
 
@@ -59,7 +79,6 @@ class StorageUtils {
       appDirectory = io.Directory(appPath);
       moviesDirectory = io.Directory(moviesPath);
 
-      // ignore: avoid_slow_async_io
       if (!await appDirectory.exists()) {
         await appDirectory.create(recursive: true);
         Utils.logInfo('[StorageUtils] - ' + 'Videos directory created');
@@ -67,12 +86,102 @@ class StorageUtils {
         Utils.logInfo('[StorageUtils] - ' + 'Videos directory already exists');
       }
 
-      // ignore: avoid_slow_async_io
       if (!await moviesDirectory.exists()) {
         await moviesDirectory.create(recursive: true);
         Utils.logInfo('[StorageUtils] - ' + 'Movies directory created');
       } else {
         Utils.logInfo('[StorageUtils] - ' + 'Movies directory already exists');
+      }
+
+      // TODO(KyleKun): synchronize async calls or something else since it seems the folder can be deleted before finishing copying files
+      // TODO(KyleKun): Elsewhere, migrate save methods to use MediaStore instead of file system
+      // Migrate old videos to new folder inside DCIM if needed
+      final io.Directory oldAppFolder = io.Directory(
+          SharedPrefsUtil.getString('appPath')
+              .replaceFirst('DCIM/OneSecondDiary/', 'OneSecondDiary/'));
+      final io.Directory oldMoviesFolder = io.Directory(
+          SharedPrefsUtil.getString('moviesPath')
+              .replaceFirst('DCIM/OneSecondDiary/Movies/', 'OSD-Movies/'));
+
+      if (await oldAppFolder.exists()) {
+        showDialog(
+          barrierDismissible: false,
+          context: Get.context!,
+          builder: (context) => AlertDialog(
+            title: const Icon(Icons.handyman),
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(10),
+            ),
+            content: Text('migrationInProgress'.tr),
+          ),
+        );
+
+        Utils.logInfo(
+          '[StorageUtils] - Migrating old videos to new folder',
+        );
+
+        try {
+          await oldAppFolder.list(recursive: false).forEach((element) async {
+            if (element is io.File && element.path.endsWith('.mp4')) {
+              final _file = SharedPrefsUtil.getString('appPath') +
+                  element.path.split('/').last;
+              await element.copy(_file);
+            }
+          });
+
+          // Copy movies folder if it exists
+          if (await oldMoviesFolder.exists()) {
+            await oldMoviesFolder
+                .list(recursive: false)
+                .forEach((element) async {
+              if (element is io.File && element.path.endsWith('.mp4')) {
+                final _file = SharedPrefsUtil.getString('moviesPath') +
+                    element.path.split('/').last;
+                await element.copy(_file);
+              }
+            });
+            await oldMoviesFolder.delete(recursive: true);
+            Utils.logWarning(
+              '[StorageUtils] - Migrated movies and deleted old folder',
+            );
+          }
+
+          await oldAppFolder.delete(recursive: true);
+          Utils.logWarning(
+            '[StorageUtils] - Migrated all videos and deleted old folder',
+          );
+          Get.back();
+          await showDialog(
+            barrierDismissible: false,
+            context: Get.context!,
+            builder: (context) => CustomDialog(
+              isDoubleAction: false,
+              title: 'success'.tr,
+              content: 'migrationSuccess'.tr,
+              actionText: 'Ok',
+              actionColor: AppColors.green,
+              action: () => Get.back(),
+            ),
+          );
+        } catch (e) {
+          Utils.logError(
+            '[StorageUtils] - Could not migrate old videos: ${e.toString()}',
+          );
+          Get.back();
+          // Just in case copying files failed
+          await showDialog(
+            barrierDismissible: false,
+            context: Get.context!,
+            builder: (context) => CustomDialog(
+              isDoubleAction: false,
+              title: 'error'.tr,
+              content: 'migrationError'.tr,
+              actionText: 'Ok',
+              actionColor: Colors.red,
+              action: () => Get.back(),
+            ),
+          );
+        }
       }
     } catch (e) {
       Utils.logError('[StorageUtils] - $e');
@@ -80,42 +189,30 @@ class StorageUtils {
   }
 
   // Create log file in internal storage
-  static Future<void> createLogFile() async {
+  static Future<void> cleanOldLogFiles() async {
     try {
-      final String appPath = SharedPrefsUtil.getString('appPath');
-      final String logPath = '$appPath/Logs/';
+      final String logsPath =
+          SharedPrefsUtil.getString('internalDirectoryPath');
 
-      // Checking if the folder really exists, if not, then create it
-      final io.Directory? logDirectory = io.Directory(logPath);
-
-      // ignore: avoid_slow_async_io
-      if (logDirectory != null && !await logDirectory.exists()) {
-        await logDirectory.create(recursive: true);
-      }
-
-      final String logFileName = Utils.getTodaysLogFilename();
-
-      // ignore: avoid_slow_async_io
-      if (!await io.File('$logPath/$logFileName').exists()) {
-        await io.File('$logPath/$logFileName').create(recursive: false);
-        SharedPrefsUtil.putString('currentLogFile', logFileName);
-      }
+      final io.Directory logsDirectory = io.Directory(logsPath);
 
       // Delete old log files (7 days past today)
-      final List<io.FileSystemEntity> files = logDirectory!.listSync();
+      final List<io.FileSystemEntity> files = logsDirectory.listSync();
       for (final io.FileSystemEntity file in files) {
-        final String filename = file.path.split('/').last;
-        final String date = filename.split('_').first;
-        final DateTime fileDate = DateTime.parse(date);
-        final DateTime today = DateTime.now();
-        final int difference = today.difference(fileDate).inDays;
+        if (file.path.endsWith('.txt') && !file.path.contains('videos')) {
+          final String filename = file.path.split('/').last;
+          final String date = filename.split('_').first;
+          final DateTime fileDate = DateTime.parse(date);
+          final DateTime today = DateTime.now();
+          final int difference = today.difference(fileDate).inDays;
 
-        if (difference > 7) {
-          await io.File(file.path).delete();
+          if (difference > 7) {
+            Utils.logInfo(
+                '[StorageUtils] - ' + 'Deleted old log file: ${file.path}');
+            await io.File(file.path).delete();
+          }
         }
       }
-
-      Utils.logInfo('[App Started] - ' + 'Log file created');
     } catch (e) {
       Utils.logError('[StorageUtils] - $e');
     }
@@ -130,7 +227,6 @@ class StorageUtils {
       // Checking if the folder really exists, if not, then create it
       final io.Directory? profileDirectory = io.Directory(profilePath);
 
-      // ignore: avoid_slow_async_io
       if (profileDirectory != null && !await profileDirectory.exists()) {
         await profileDirectory.create(recursive: true);
       }
